Add support for 4-byte wchar_t (#42)
Some of our target platforms use four bytes for wchar_t, whereas others
use two. This change adds a class that is templated on sizeof(wchar_t)
in order to pick the correct conversion based on the size; to/from UTF-16
for cases where wchar_t is two bytes, and to/from UCS-4/UTF-32 for cases
where wchar_t is four bytes.

Note: This change makes the name to_utf16string problematic, so it has
been renamed to to_widestring. This does not feel ideal, as it doesn't
line up with the to_utf8string name, but renaming that to to_string felt
odd. A previous project I worked on used to_utf8string and from_utf8string.
I like those names, but it would only make sense to use them if we removed
the overload that takes a std::wstring as its parameter.