Fix potential race during cancel and running async call (#278)
There is a potential narrow race condition in async lib if you cancel an async call while that call is in flight. This can cause the completion callback to be invoked, which is allowed to delete the async block. If an async call was running at the time, it's provider data structure now contains a dangling pointer.

This was solved earlier by introducing a "zombie block" which is a copy of the user's async block that we swapped in when we were cleaning up. This can cause the async field of async provider data to switch to a different pointer value, and while narrow, there is still a potential for a race with a running call.

This change keeps a local copy of the user's async block. It promotes the "zombie block" idea such that the async provider data always gets the copy, while user callbacks always get the original pointer. We keep state transitions of the two in lockstep always. This way no matter what the user does with the async block the async provider method always has valid data.