Abstract async queue (#291)
This PR primarily moves async queue away from std::mutex and instead uses spin locks, at least on multi-processor systems. The reason for this is to prevent context switches while adding things to the queue, which can slow down perf critical threads.

The async queue implementation is also now built on an interface abstraction, which will make interoperability between Windows code and libHttpClient easier.

Finally, there are some API additions / changes:

DuplicateAsyncQueueHandle now returns a HRESULT and has the new handle as an out-param. The semantics haven't changed, but this signature mates better with other Win32 "duplicate" APIs and allows us a little more freedom in the implementation later on.

There are two new dispatch modes: SerializedThreadPool uses the thread pool, but ensures that queue calls are run in series through it instead of in parallel. This is useful if you want to use a thread pool as your completion dispatch mode, but still want things to always come in order. Immediate dispatch mode doesn't queue at all and just invokes the callback immediately on the submitting thread. This is a useful utility mechanism for chaining async calls together.

RemoveAsyncQueueCallbacks can now accept a nullptr as the callback function pointer to act as a wild card. This must be used with caution, since you can remove any callback (not just yours), but it's kind of necessary to support lambdas.