Fix a subtle race shutting down a task queue with deferred calls (#464)
If a task queue is executing a deferred callback and is asked to terminate, it's possible the the deferred call runs while the queue task port is destructing. This can cause a crash.  

This also changes cleanup so if you don't have a payload to return your cleanup op is called at the end of work processing.  this can help with shutdown.