Sync from OS tree (#538)
There was a minor bug fix in the Windows OS tree to clear the header data of a task queue before it is deleted so if someone tried to use a released task queue handle we could catch it early. If you reused right after releasing the un-allocated memory could still look like a valid task queue and this will cause crashes later on.