Fix for possible spin during queue termination (#539)
If a task queue is terminated while tasks are still being dispatched, the logic that drains the queue before termination can get into an infinite loop. The drain logic needs to pop items off the queue and those that should not be removed are re-added. The first one to be re-added is remembered as a sentinel so the loop knows when to quit. If that item gets dispatched, the loop never terminates.

With this change we put all the items to re-add to the queue into a local buffer and then push them back onto the queue later.