Fix possible race condition during cancel (#557)
Co-authored-by: Jason Sandlin <jasonsa@microsoft.com>

If a call completes with either a zero byte payload or an error, the async lib will try to call Cleanup on the call provider early, instead of the normal cleanup which occurs after the completion event has been fired. This early cleanup can create a race condition with XAsyncCancel as follows:

XAsyncCancel is called. Early in the method it acquires a reference on the state structure.
The provider completes its DoWork with zero payload or error and rolls out of the method.
Early cleanup of the provider now happens.
XAsyncCancel continues and tries to call Cancel on the provider.
The provider is likely to crash now because it has already run its cleanup code.
This race was observed with the networking stack on Xbox.

My fix for this is to introduce another "cleanup location" and to change the code that allows cleanup after DoWork to only set this value if the current location is the default Destructor setting. The net effect of this is if the race occurs the async provider cleanup is deferred to state cleanup and loses out on the early cleanup optimization. This seemed better than ditching the early cleanup route entirely.