Fix some faulty logic in WinHttp WebSocket fragment handling (#670)
When WebSocket message fragments are received, they are collected in a buffer within LHC.  They are only forwarded to the client when that buffer is full or when the final fragment has been received.  When WinHttp has given us fragments, LHC incorrectly invokes the HCWebSocketBinaryMessageFragmentFunction even if we were able to accumulate the fragments in our buffer before passing them along to the client.

This also fixes a bug where the WebSocket's default m_maxReceiveBufferSize isn't correctly initialized to 20kb as documented in the header.