Cherry-pick WinHttp WebSocket fragment changes (#674) (#675)
* Proper handling of large incoming messages for WinHttp WebSockets  (#653)

* Fixes bug where messages that exceeded the hardcoded receive buffer size were never delivered to clients
* Added public API to set a message fragment handler. When messages are broken down at the platform level and exceed the configured receive buffer size, they will be passed to clients in fragments.
* Adds public API to configure maximum WebSocket receive buffer size.  When receive buffer is full, partial messages are passed to client.
* Adds public API to configure the maximum receive buffer size (WinHttp only).  If large messages are expected, this can be changed so that full messages are delivered all at once rather than in chunks.  Setting this buffer size to large values does mean that the entire message will be stored in LHC memory until it is passed along to clients so memory performance may be affected.

* Fix winHttp WebSocket regression cause by message fragment changes (#663)

* Fix some faulty logic in WinHttp WebSocket fragment handling (#670)

When WebSocket message fragments are received, they are collected in a buffer within LHC.  They are only forwarded to the client when that buffer is full or when the final fragment has been received.  When WinHttp has given us fragments, LHC incorrectly invokes the HCWebSocketBinaryMessageFragmentFunction even if we were able to accumulate the fragments in our buffer before passing them along to the client.

This also fixes a bug where the WebSocket's default m_maxReceiveBufferSize isn't correctly initialized to 20kb as documented in the header.