Re-classifying NO_NETWORK to be only UnknownHostException on Android (#709)
* Re-classifying NO_NETWORK to be only UnknownHostException
At some point recently various socket exceptions started getting classified as "isNoNetworkFailure". While this was intended to be helpful classification, the resulting behavior is that on a socket exception, retry logic will not get engaged. This is because retry logic does not attempt to retry if there is no network in order to save from repeatedly waiting when there isn't a reason to wait. This behavior change was not good or intended so I am reverting it.