Sometimes XTaskQueueTerminate does not signal the queue is terminated. (#858)
Sometimes XTaskQueueTerminate does not actually terminate the queue. If you wait for this to complete it never does.

The problem turned out to be how we are using lockfreequeue::remove_if. This API works by popping all items off the queue, invoking a callback, and items that the callback returned false for are placed on a temporary queue and later added back.

If two threads are running at the same time, it is possible for T2 to hide items T1 wants, and vice versa.

The fix for this is just to spinlock in remove_if. The duration of time spent inside this is very small, and contention is very rare.

This change also picks up changes form the OS repo that we didn't propagate. The OS code moved to a different HRESULT for bad handles and a "runtime iteration" that allows the runtime to be reloaded.